% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{cohens_d}
\alias{cohens_d}
\title{Effect size using Hedges' \emph{g}}
\usage{
cohens_d(mu_x, sd_x, n_x, mu_y = NULL, sd_y = NULL, n_y = NULL,
  one_sample_test_val = 0)
}
\description{
Effect size using Hedges' \emph{g}
}
