% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dtbl2list}
\alias{dtbl2list}
\title{Split a data.table into separate lists by group}
\usage{
dtbl2list(data, ...)
}
\arguments{
\item{data}{a data.frame or data.table}

\item{...}{unquoted column names}
}
\value{
a list of data.table/data.frame objects
}
\description{
Split a data.table into separate lists by group
}
\examples{
data <- cars
dtbl2list(data, speed)
}
