% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{example_stan}
\alias{example_stan}
\title{Example stan model and data}
\usage{
example_stan(n = 30, seed = NULL, data_only = FALSE)
}
\arguments{
\item{n}{sample size per group}

\item{seed}{seed value to get same data samples}

\item{data_only}{return only the data without fitting the model}
}
\value{
list(data, pars, inits, model, fit)
}
\description{
Example stan model and data
}
\examples{
stan_ex <- example_stan(seed=19851905)
}
