% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{example_stanreg}
\alias{example_stanreg}
\title{Stan example model}
\usage{
example_stanreg(iter = 1000, chains = 2)
}
\arguments{
\item{iter}{number of iterations for stan_glmer}

\item{chains}{number of chains for stan_glmer}
}
\value{
stanfit obj
}
\description{
Stan example model
}
\examples{
stanfit <- example_stanreg()
}
