% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_posterior.R
\name{geom_posterior}
\alias{geom_posterior}
\title{Geom for plotting posterior distributions}
\usage{
geom_posterior(mapping = NULL, data = NULL, stat = "posterior_density",
  position = "spread", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{Use to override the default connection between
\code{geom_violin} and \code{stat_ydensity}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{center_stat}{character string of function to compute center of distribution, such as \code{"median"} or \code{"mean"}}

\item{interval_width}{width of the confidence interval, e.g., 0.95}

\item{interval_type}{method of computing the interval, either \code{"hdi"} or \code{"ci"}}
}
\description{
Geom for plotting posterior distributions
}
\examples{
df <-
data.frame(
l = rep(letters[1:4], each = 2500),
v = rnorm(10000),
f = rep(LETTERS[1:2], each = 5000)
)
ggplot(df) +
geom_posterior(
aes(y = f, x = v, group = l),
interval_width = 0.95,
interval_type = 'hdi',
center_stat = 'median')
}
