% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{loo_table}
\alias{loo_table}
\title{table of LOO comparisons}
\usage{
loo_table(loo_list, stat = "looic")
}
\arguments{
\item{loo_list}{a list of objects of the type \code{loo}}

\item{stat}{}
}
\value{
data.table of loo comparisons
}
\description{
table of LOO comparisons
}
\examples{
stanreg <- example_stanreg()
stanreg2 <- update(stanreg, . ~ . + floor:log_uranium)
loo_list <- lapply(list(stanreg, stanreg2), rstanarm::loo)
loo_table(loo_list)
}
