% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_data_and_posterior}
\alias{merge_data_and_posterior}
\title{Merge posterior predictions with data}
\usage{
merge_data_and_posterior(stanreg, newdata = NULL,
  post_fun = rstanarm::posterior_predict, var.sample = ".sample",
  var.rows = ".obs", var.yhat = ".y", ...)
}
\arguments{
\item{newdata}{data to use for prediction}

\item{...}{options passed to rstanarm::posterior_predict}
}
\value{
\link[data.table:data.table]{data.table::data.table}
}
\description{
Merge posterior predictions with data
}
\examples{
#TODO:
}
