% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{n_mcmc_samples}
\alias{n_mcmc_samples}
\title{Number of MCMC samples from model}
\usage{
n_mcmc_samples(object)
}
\arguments{
\item{object}{stanreg or stanfit object}
}
\value{
scalar value
}
\description{
Number of MCMC samples from model
}
\examples{

}
