% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nlist}
\alias{nlist}
\title{Create a list from object names}
\usage{
nlist(...)
}
\arguments{
\item{...}{same as in a regular list}
}
\description{
Create a list from object names
}
\examples{
x <- 5
y <- 'stuff'
nlist(x, y)
}
