% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{one_sample_mvt_make}
\alias{one_sample_mvt_make}
\title{One sample multivariate-t}
\usage{
one_sample_mvt_make(pars, sample_size_list = NULL)
}
\arguments{
\item{pars}{list of parameters needed for simulation and to create objects to pass to stan}
}
\value{
sim_fit_template
}
\description{
Simulation function for one sample multivariate t
}
