% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{pairwise_contrasts}
\alias{pairwise_contrasts}
\title{posterior predictive contrasts for all pairwise comparisons in a list}
\usage{
pairwise_contrasts(stanreg, cdata, width = 0.95, draws = NULL,
  yfun = median)
}
\arguments{
\item{stanreg}{stanreg object from rstanarm.}

\item{cdata}{a list of data sets to use to find posterior predictions of
each.}

\item{width}{confidence interval width. Sent to the \code{prob} argument
from \code{\link[rstanarm]{posterior_interval}}.}

\item{draws}{number of samples to use from the posterior predictive
distribution. Sent to the \code{draws} argument from
\code{\link[rstanarm]{posterior_interval}}.}

\item{yfun}{the stat function to use to collapse the predictions into a
scalar value. Corresponds to the mean of the predicted responses for each
data set in \code{cdata}.}
}
\value{
list containing CI matrix and posterior differences
}
\description{
posterior predictive contrasts for all pairwise comparisons in a list
}
\examples{
stanreg <- example_stanreg()
cdata <- list(
 c1 = contrast_data(stanreg, TRUE, subset_expression = .~wt < 2),
 c2 = contrast_data(stanreg, TRUE, subset_expression = .~wt < 2.75),
 c3 = contrast_data(stanreg, TRUE, subset_expression = .~wt > 3))
pairwise_contrasts(stanreg, cdata)$ci
}
