% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{post_pred_dtable}
\alias{post_pred_dtable}
\title{Posterior predict data.table}
\usage{
post_pred_dtable(object, newdata = NULL, append = TRUE,
  fn = rstanarm::posterior_predict, ...)
}
\arguments{
\item{object}{stanfit object}

\item{newdata}{data to use for prediction}

\item{append}{add data columns to predictions}

\item{fn}{posterior_predict or posterior_linpred}

\item{...}{options passed to rstanarm::posterior_predict}
}
\value{
Posterior prediction matrix as a data.table, merging newdata if available.
}
\description{
Posterior predict data.table
}
\examples{
#TODO:
}
