% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{pp_contrast}
\alias{pp_contrast}
\title{Posterior predictive contrast}
\usage{
pp_contrast(stanreg, cdata, ccoef, width = 0.95, yfun = median, ...)
}
\arguments{
\item{stanreg}{stanreg object from rstanarm.}

\item{cdata}{a list of data sets to use to find posterior predictions of
each.}

\item{ccoef}{contrast coefficients corresponding to the order of the data
sets in \code{cdata}.}

\item{width}{confidence interval width. Sent to the \code{prob} argument
from \code{\link[rstanarm]{posterior_interval}}.}

\item{yfun}{the stat function to use to collapse the predictions into a
scalar value. Corresponds to the average of the predicted responses for each
data set in \code{cdata}. Default is median.}

\item{...}{additional parameters passed to the function \code{\link[rstanarm]{posterior_predict}}
\code{\link[rstanarm]{posterior_interval}}.}
}
\value{
list containing the confidence interval and posterior predictive
  contrast difference
}
\description{
Posterior predictive contrast
}
\examples{
stanreg <- example_stanreg()
cdata <- list(
 c1 = contrast_data(stanreg, TRUE, margin_ignore = 'floor', subset_expression = .~ log_uranium < -0.5),
 c2 = contrast_data(stanreg, TRUE, margin_ignore = 'floor', subset_expression = .~ log_uranium >= -0.5 & log_uranium <= 0.5),
 c3 = contrast_data(stanreg, TRUE, margin_ignore = 'floor', subset_expression = .~ log_uranium > 0.5))

my_contrast <- pp_contrast(stanreg, cdata, c(-1, 0.5, 0.5))
my_contrast$ci
}
