% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_stanarm}
\alias{print_stanarm}
\title{Alternative rstanarm print}
\usage{
print_stanarm(stanreg, header = "Stan model summary")
}
\arguments{
\item{stanreg}{a stanreg object}

\item{header}{header name}
}
\description{
Alternative rstanarm print
}
