% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pval_format}
\alias{pval_format}
\title{string format p-value cutoffs}
\usage{
pval_format(p)
}
\arguments{
\item{p}{a p value between 0 and 1}
}
\value{
character matrix
}
\description{
string format p-value cutoffs
}
\examples{
pval_format(.055)
pval_format(.05)
pval_format(.049)
pval_format(.01)
pval_format(.001)
pval_format(.0001)

p <- seq(0, 0.06, .01)
data.frame(p, pval_format(p))
}
