\name{rbaes-package}
\alias{rbaes-package}
\alias{rbaes}
\docType{package}
\title{
\packageTitle{rbaes}
}
\description{
\packageDescription{rbaes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rbaes}
\packageIndices{rbaes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rbaes}

Maintainer: \packageMaintainer{rbaes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
