% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rev_unit_scale}
\alias{rev_unit_scale}
\title{Reverse unit scaling}
\usage{
rev_unit_scale(x, y = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{object to be scaled back to original scale}

\item{y}{optional object that has center, scale, and adj attributes to use on x. If y is not
used, these attributes will be taken from x.}
}
\value{
numeric vector in original scale
}
\description{
Reverse unit scaling
}
\examples{
z <- rnorm(10000, 100, 15)
y <- unit_scale(z)
rev_unit_scale(y)
}
