% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{stan_chain_convergence}
\alias{stan_chain_convergence}
\title{Chain convergence stats}
\usage{
stan_chain_convergence(stanmodel, pars, view = FALSE)
}
\arguments{
\item{stanmodel}{A fitted stan model}

\item{pars}{Names of parameters to calculate convergence statistics, defaults to all parameters, including lp__}

\item{view}{Plot each chain's log-prob and its distribution}
}
\description{
This is to see which chains should be removed, if any.
}
\details{
If you get an error while plotting and using RStudio, try to make plot window bigger.
}
\examples{
rstan_pack <- stan_mejr(parallel=TRUE)
chain_convergence(rstan_pack$stan_mcmc)
}
