% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stan_chol}
\alias{stan_chol}
\title{Stan formatted Cholesky factored cov/cor matrix}
\usage{
stan_chol(mat)
}
\arguments{
\item{mat}{A cholesky factored covariance or correlation matrix}
}
\description{
Stan formatted Cholesky factored cov/cor matrix
}
\examples{
# make matrix
L <- solve(rWishart(1, 6, diag(5))[,,1])

# compare
l1 <- chol(L)
l2 <- stan_chol(L)
}
