% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{stan_drop_chains}
\alias{stan_drop_chains}
\title{Drop unwanted chains from stanfit object}
\usage{
stan_drop_chains(object, chain_ids)
}
\arguments{
\item{object}{object of class 'stanfit'}

\item{chain_ids}{an integer vector chain numbers to drop}
}
\value{
new instance of stanfit class
}
\description{
Drop unwanted chains from stanfit object
}
\examples{
stan_drop_chains(myFittedObject, c(2,4))
}
