% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{stan_mejr}
\alias{stan_mejr}
\title{Stan Mejr}
\usage{
stan_mejr(model, name, data, pars, samples, init, out = NULL,
  parallel = FALSE, debug = FALSE, repackage = NULL, ...)
}
\arguments{
\item{model}{model code or file name, both as character objects}

\item{name}{name of model being fitted}

\item{data}{data in stan format}

\item{pars}{named parameters. Defaults to kept parameters}

\item{samples}{list of sample parameters. See example for list structure.}

\item{init}{initial values}

\item{out}{directory for where to save output files}

\item{parallel}{use multiple cores for running multiple chains}

\item{debug}{do a test run with options saved to global env}

\item{repackage}{add additional objects to the output of this function}

\item{...}{other arguments from the main \code{rstan::stan} function}

\item{control}{control paramters for algorithm}
}
\value{
list of objects: \code{stan_mcmc, pram, central, env, repack}
}
\description{
Wrapper function for simplifying \code{stan}. See \code{?rstan::stan} for more info.
}
\details{
Also provides additional plots and repackages data into a single object
}
\examples{
# Using fake data for example
stan_dat <- stan_test_data()

# fit model
stan_fit <- stan_mejr(
     model=stan_dat$model,
     name="example_model",
     data=stan_dat$data,
     pars=stan_dat$pars,
     samples=list(n_chains = 4, n_final = 1200, n_thin = 2, n_warm = 800),
     init=stan_dat$init,
     out="~/Desktop",
     parallel=FALSE)

# Run default example model
stan_fit <- rstan_debug <- stan_mejr(debug=TRUE)
}
