% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{stan_plots_mejr}
\alias{stan_plots_mejr}
\title{RStan plots}
\usage{
stan_plots_mejr(stan_obj, pars, out, label, inc_warmup = FALSE,
  max_pars = 9)
}
\arguments{
\item{stan_obj}{The fitted object from \code{rstan::stan}}

\item{pars}{A character vector of parameter names to plot. Defaults to working directory.}

\item{out}{Path for where to save the plots}

\item{label}{a character string to attach to file names}

\item{inc_warmup}{Include warmup for all plots. Defaults to \code{FALSE}}
}
\value{
nothing. Prints plots to disk
}
\description{
Saves a series of plots from a fitted rstan object
}
\examples{
stan_fit <- stan_mejr()
stan_plots_mejr(stan_fit$stan_mcmc)
}
