% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{stan_point_est}
\alias{stan_point_est}
\title{STAN point estimates}
\usage{
stan_point_est(stan_obj, ...)
}
\arguments{
\item{stan_obj}{Stan fitted object}

\item{...}{additional options passed to function \code{hdiq}}
}
\value{
list object with same parameter names
}
\description{
STAN point estimates
}
