% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{stan_test_data}
\alias{stan_test_data}
\title{Stan test data}
\usage{
stan_test_data(n = 100)
}
\arguments{
\item{n}{number of observations}
}
\value{
list of stan options
}
\description{
Stan test data
}
\examples{
fake_data <- stan_test_data(n=30)
}
