% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{stan_yhat}
\alias{stan_yhat}
\title{Predict values from X matrix and posterior samples}
\usage{
stan_yhat(effects_list, link = FALSE, link_fun = identity)
}
\arguments{
\item{effects_list}{A list of named lists that each contain an X matrix and the parameters values B}

\item{link}{Logical value deciding whether to link each batch in a single data.table or keep as a list}

\item{link_fun}{The linking function after summing predicted values. Defaults to the identity function}
}
\value{
A list or data.table
}
\description{
Predict values from X matrix and posterior samples
}
\examples{
# Two batches of effects that will be summed, and
# the exponential link function used afterwards.
# Each batch has an nObservations x nParameters model matrix and
#  an nSamples x nParameters posterior matrix
effects_list <- list(
    main_fx  = list(X = model.matrix(~X1*X2, data = test_data),
                    B = posterior_samples$Beta),
    other_fx = list(X = model.matrix(~0 + grp, data = test_data),
                    B = posterior_samples$Gamma)

predicted_values <- stan_yhat(effects_list, TRUE, exp)

# Numerical example:
effects_list <- list(
    fx = list(
        X = array(runif(25), c(5, 5)),
        B = array(runif(1000), c(200, 5))
    )
)
Y <- stan_yhat(effects_list, TRUE, sigmoid)
}
