% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{stanreg_dtbl}
\alias{stanreg_dtbl}
\title{Get model data from stanreg}
\usage{
stanreg_dtbl(stanreg, get_y = TRUE)
}
\arguments{
\item{stanreg}{object of class stanreg}

\item{get_y}{logical value. Include the output columns?}
}
\value{
data.table
}
\description{
Get model data from stanreg
}
\examples{
stanreg <- example_stanreg()
model_data <- stanreg_dtbl(stanreg, TRUE)
}
