% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unit_scale}
\alias{unit_scale}
\title{Scale a numeric vector}
\usage{
unit_scale(x, y = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{y}{alternative numeric vector that contains the attributes to use to rescale x}

\item{na.rm}{remove NAs from the vector before finding stats}
}
\value{
scaled vector
}
\description{
Vectors are scaled to have a mean of 0.5 and standard deviation of 0.5.
}
\details{
The function works similar to scale, in that attributes of the original mean and standard
deviation are saved in the object called \code{center} and \code{scale}. The adjustment is also
provided in the \code{adj} attribute.
}
\examples{
x <- rnorm(1000, 50, 5)
hist(x, br=50)

y <- unit_scale(x)
hist(y, br=50)
mean(y) # <- 0.5
sd(y) # <- 0.5
}
