% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{pairwise_contrasts}
\alias{pairwise_contrasts}
\title{posterior predictive contrasts for all pairwise comparisons in a list}
\usage{
pairwise_contrasts(stanreg, cdata, width = 0.95, yfun = median,
  pp_all = FALSE, ...)
}
\arguments{
\item{stanreg}{stanreg object from rstanarm.}

\item{cdata}{a list of data sets to use to find posterior predictions of
each.}

\item{width}{confidence interval width. Sent to the \code{prob} argument
from \code{\link[rstanarm]{posterior_interval}}.}

\item{yfun}{the stat function to use to collapse the predictions into a
scalar value. Corresponds to the average of the predicted responses for each
data set in \code{cdata}. Default is median.}

\item{pp_all}{posterior predictions for each list item in cdata. Defaults to
\code{FALSE}}

\item{...}{additional parameters passed to the function \code{\link[rstanarm]{posterior_predict}}
\code{\link[rstanarm]{posterior_interval}}.}
}
\value{
list containing CI matrix and posterior differences
}
\description{
posterior predictive contrasts for all pairwise comparisons in a list
}
\examples{
stanreg <- example_stanreg()
cdata <- list(
c1 = contrast_data(stanreg, TRUE, margin_ignore = 'floor', subset_expression = .~ log_uranium < -0.5),
c2 = contrast_data(stanreg, TRUE, margin_ignore = 'floor', subset_expression = .~ log_uranium >= -0.5 & log_uranium <= 0.5),
c3 = contrast_data(stanreg, TRUE, margin_ignore = 'floor', subset_expression = .~ log_uranium > 0.5))
pairwise_contrasts(stanreg, cdata)$ci.cont
}
