% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pooled_sd}
\alias{pooled_sd}
\title{Hedges G}
\usage{
pooled_sd(sd, n)
}
\arguments{
\item{n_vec}{sample sizes corresponding to the standard deviations}
}
\value{
numeric value
}
\description{
Hedges G
}
\examples{
pooled_sd(c(1, 5), c(100, 8))
}
