% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{pp_contrast}
\alias{pp_contrast}
\title{Posterior predictive contrast}
\usage{
pp_contrast(stanreg, cdata, ccoef, width = 0.95, draws = NULL,
  yfun = median)
}
\arguments{
\item{stanreg}{stanreg object from rstanarm.}

\item{cdata}{a list of data sets to use to find posterior predictions of
each.}

\item{ccoef}{contrast coefficients corresponding to the order of the data
sets in \code{cdata}.}

\item{width}{confidence interval width. Sent to the \code{prob} argument
from \code{\link[rstanarm]{posterior_interval}}.}

\item{draws}{number of samples to use from the posterior predictive
distribution. Sent to the \code{draws} argument from
\code{\link[rstanarm]{posterior_interval}}.}

\item{yfun}{the stat function to use to collapse the predictions into a
scalar value. Corresponds to the mean of the predicted responses for each
data set in \code{cdata}.}
}
\value{
list containing the confidence interval and posterior predictive
  contrast difference
}
\description{
Posterior predictive contrast
}
\examples{
stanreg <- example_stanreg()
 cdata <- list(
   c1 = contrast_data(stanreg, TRUE, subset_expression = .~wt < 2),
   c2 = contrast_data(stanreg, TRUE, subset_expression = .~wt < 2.75),
   c3 = contrast_data(stanreg, TRUE, subset_expression = .~wt > 3))
pp_contrast(stanreg, cdata, c(-1, 0.5, 0.5))
}
