% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbaes-package.R
\docType{package}
\name{rbaes-package}
\alias{rbaes-package}
\title{R Bayesian Analysis and Estimation Stuff}
\description{
A set of functions and routines for Bayesian data analysis.
}
\details{
Mostly wrappers and stuff, or simplifying some operations.
}
\seealso{
\link[rstan:stan]{rstan::stan}, \link[rstanarm:stan_glmer]{rstanarm::stan_glmer}

See \code{help(package = "rbaes")} for a list of functions.

View vignettes with \code{browseVignettes(package ="rbaes")}.

Below are some useful links:
\itemize{
\item http://mc-stan.org/
}
}
