% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbaes-package.R
\docType{package}
\name{rbaes}
\alias{rbaes}
\alias{rbaes-package}
\title{R Bayesian Analysis and Estimation Stuff}
\description{
A set of functions and routines for Bayesian data analysis
}
\section{Info}{
 See \code{library(help="rbaes")} for help. \cr \cr Below are
some useful links I can click on instead of google searching them: \cr \cr
\url{http://mc-stan.org/}
}

