% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stanreg_dtbl}
\alias{stanreg_dtbl}
\title{Get model data from stanreg}
\usage{
stanreg_dtbl(stanreg, model_frame = FALSE, get_y = FALSE)
}
\arguments{
\item{stanreg}{object of class stanreg}

\item{model_frame}{If \code{TRUE}, then extract from \code{$glmod$fr}, else use
\code{$data}.}

\item{get_y}{logical value. Include the output columns?}
}
\value{
data.table
}
\description{
Get model data from stanreg
}
\examples{
stanreg <- example_stanreg()
model_data <- stanreg_dtbl(stanreg, TRUE, TRUE)
}
